+++ test/send.js
@@ -853,56 +853,6 @@ describe('send(file).pipe(res)', function () {
     })
   })
 
-  describe('.maxage()', function () {
-    it('should default to 0', function (done) {
-      var app = http.createServer(function (req, res) {
-        send(req, 'test/fixtures/name.txt')
-          .maxage(undefined)
-          .pipe(res)
-      })
-
-      request(app)
-        .get('/name.txt')
-        .expect('Cache-Control', 'public, max-age=0', done)
-    })
-
-    it('should floor to integer', function (done) {
-      var app = http.createServer(function (req, res) {
-        send(req, 'test/fixtures/name.txt')
-          .maxage(1234)
-          .pipe(res)
-      })
-
-      request(app)
-        .get('/name.txt')
-        .expect('Cache-Control', 'public, max-age=1', done)
-    })
-
-    it('should accept string', function (done) {
-      var app = http.createServer(function (req, res) {
-        send(req, 'test/fixtures/name.txt')
-          .maxage('30d')
-          .pipe(res)
-      })
-
-      request(app)
-        .get('/name.txt')
-        .expect('Cache-Control', 'public, max-age=2592000', done)
-    })
-
-    it('should max at 1 year', function (done) {
-      var app = http.createServer(function (req, res) {
-        send(req, 'test/fixtures/name.txt')
-          .maxage(Infinity)
-          .pipe(res)
-      })
-
-      request(app)
-        .get('/name.txt')
-        .expect('Cache-Control', 'public, max-age=31536000', done)
-    })
-  })
-
   describe('.root()', function () {
     it('should set root', function (done) {
       var app = http.createServer(function (req, res) {